Seperate Stats Parser classes into their own files
I am planning on adding unit tests for these parsers. Seperating them
out should make this easier, it keeps the files smaller and more
readable, plus it keeps things more organised.
